feat: deixando icones e rotas sendo passadas como prop
